Make throttling compile time configurable
