+++ internal/throttle.go
@@ -1,10 +1,7 @@
 package sslr
 
 import (
-	"math"
 	"time"
-
-	"github.com/erkkah/letarette/pkg/logger"
 )
 
 type throttledOperation struct {
@@ -13,40 +10,4 @@ type throttledOperation struct {
 	startTime        time.Time
 	totalJobDuration time.Duration
 	jobStartTime     time.Time
-}
-
-func newThrottle(name string, percentage float64) *throttledOperation {
-	return &throttledOperation{
-		name:  name,
-		level: math.Max(1, math.Min(percentage, 100)) / 100,
-	}
-}
-
-func (t *throttledOperation) start() {
-	logger.Debug.Printf("Started %s", t.name)
-	t.jobStartTime = time.Now()
-	if t.startTime.IsZero() {
-		t.startTime = t.jobStartTime
-	}
-}
-
-func (t *throttledOperation) end() {
-	logger.Debug.Printf("Ended %s", t.name)
-	t.totalJobDuration += time.Since(t.jobStartTime)
-}
-
-func (t *throttledOperation) wait() {
-	totalDuration := time.Since(t.startTime)
-	utilization := float64(t.totalJobDuration.Milliseconds())
-	logger.Debug.Printf("Utilization %.2f%%", 100*utilization/float64(totalDuration.Milliseconds()))
-	if t.level >= 1 {
-		return
-	}
-
-	utilizationLimit := float64(totalDuration.Milliseconds()) * t.level
-	if utilization > utilizationLimit {
-		waitTime := time.Duration(2*(utilization-utilizationLimit)) * time.Millisecond
-		logger.Debug.Printf("Waiting %v to keep utilization at %.2f%%", waitTime, t.level*100)
-		time.Sleep(waitTime)
-	}
 }
